+++ router.py
@@ -1,0 +1,5 @@
+from services.inventory_service import InventoryManagingService
+from services.order_processing_service import OrderProcessingService
+
+class Router(object):
+    pass
