+++ utils/database_provider.py
@@ -1,4 +1,5 @@
 from typing import Any
+
 import sqlalchemy as sa
 from sqlalchemy.exc import ResourceClosedError
 import pandas as pd
